Define `crate::shared::pick_module` helper
