+++ src/commands/list.rs
@@ -44,73 +44,7 @@ pub async fn list(
             EX_USAGE
         })?;
 
-        let module = if let Some(want) = module {
-            let module = modules.iter().find(|m| m.name == want).ok_or_else(|| {
-                ceprintln!("<s,r>error:</> failed to find a module named `{want}` that supports cataloging the URL: `{input_url}`");
-                EX_SOFTWARE
-            })?;
-
-            if registry
-                .is_module_enabled(&module.name)
-                .await
-                .map_err(|e| {
-                    ceprintln!(
-                        "<s,r>error:</> error while checking whether module `{}` is enabled: {e}",
-                        module.name
-                    );
-                    EX_IOERR
-                })?
-            {
-                module
-            } else {
-                ceprintln!(
-                    "<s,r>error:</> module <s>{}</> is not enabled.",
-                    module.name
-                );
-                ceprintln!(
-                    "<s,dim>hint:</> It can be enabled with: `asimov module enable {}`",
-                    module.name
-                );
-                return Err(EX_UNAVAILABLE);
-            }
-        } else {
-            let mut iter = modules.iter();
-            loop {
-                let module = iter.next().ok_or_else(|| {
-                    ceprintln!(
-                        "<s,r>error:</> failed to find a module to catalog the URL: `{input_url}`"
-                    );
-                    let module_count = modules.len();
-                    if module_count > 0 {
-                        if module_count == 1 {
-                            ceprintln!("<s,dim>hint:</> Found <s>{module_count}</> installed module that could handle this URL but is disabled.");
-                        } else {
-                            ceprintln!("<s,dim>hint:</> Found <s>{module_count}</> installed modules that could handle this URL but are disabled.");
-                        }
-                        ceprintln!("<s,dim>hint:</> A module can be enabled with: `asimov module enable <<module>>`");
-                        ceprintln!("<s,dim>hint:</> Available modules:");
-                        for module in &modules {
-                            ceprintln!("<s,dim>hint:</>\t<s>{}</>", module.name);
-                        }
-                    }
-                    EX_UNAVAILABLE
-                })?;
-
-                if registry
-                    .is_module_enabled(&module.name)
-                    .await
-                    .map_err(|e| {
-                        ceprintln!(
-                            "<s,r>error:</> error while checking whether module `{}` is enabled: {e}",
-                            module.name
-                        );
-                        EX_IOERR
-                    })?
-                {
-                    break module;
-                }
-            }
-        };
+        let module = shared::pick_module(&registry, &input_url, modules.as_slice(), module).await?;
 
         let mut cataloger = asimov_runner::Cataloger::new(
             format!("asimov-{}-cataloger", module.name),
